Created a new agent class
Created an agent class so that we can pass in the SSH client to define a tool that the agent can use to run commands in the Pwnbox instance. Added  two new lines to `darkcircuit_app_win.py` that are commented out but show how the new class could be used.