Adjustments to exe
fixed bugs with exe and terminal stuff now that its all running without modal. Because of not running in Modal the Terminal window communication between the agent is much cleaner and doesn't use UI anymore. If there is time I will clean up all of the old UI terminal comms that are not needed.